Move base attributes and validation to WeatherBaseForm
